fixed inconsistent backprop

+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/transforms/HardTanh.java
@@ -40,7 +40,7 @@ public class HardTanh extends BaseElementWiseOp {
             if(ret < -1)
                 return -1;
             else if(ret > 1)
-                return 1;
+                return (float)  1;
             else
                 return ret;
         }
